Move environment_logs into the existing container_logs.rs
Step 6 (final) of splitting containers.rs. environment_logs only
depended on the already-pub(crate) stack_directory/runtime_command/
detect_runtime, so it moves cleanly next to the log-streaming spawn()
already in container_logs.rs, re-exported as `pub use
crate::container_logs::{environment_logs, LogProcess};` for
runtime_commands.rs.

Narrower than planned: clear_service_logs stays in containers.rs — it
depends on the private service_context() helper shared by
service_configuration/execute_service_command/service_url, so moving
it would have meant loosening that helper's visibility for one
external caller, the same tradeoff that kept service_url and
refresh_site_routes in containers.rs in the previous step.

containers.rs is now 1543 lines, down from 3446 at the start of this
refactor (container_schema.rs 297, container_validation.rs 460,
container_compose.rs 770, container_bootstrap.rs 348,
container_gateway.rs 385, container_logs.rs 101 hold the rest) — an
orchestrator for CRUD/lifecycle/inspection, with schema, validation,
compose/Dockerfile generation, database bootstrap, gateway
orchestration, and log streaming each in their own focused module,
following the container_<concern>.rs convention already established
by container_inspection.rs/container_lifecycle.rs/container_routes.rs/
container_runtime.rs.
