+++ src-tauri/src/container_logs.rs
@@ -62,3 +62,40 @@ pub fn spawn(
         secrets: crate::security::environment_secrets(app, id),
     })
 }
+
+pub fn environment_logs(app: &tauri::AppHandle, id: &str) -> Result<String, String> {
+    let directory = crate::containers::stack_directory(app, id)?;
+    if !directory.join("compose.yaml").exists() {
+        return Ok(String::new());
+    }
+    let preferred = crate::settings::load(app)?.map(|settings| settings.runtime);
+    let runtime = crate::containers::detect_runtime(preferred.as_deref());
+    let executable = runtime
+        .runtime
+        .filter(|_| runtime.running && runtime.compose_available)
+        .ok_or(runtime.message)?;
+    let output = crate::process::output(
+        crate::containers::runtime_command(&executable)
+            .args(["compose", "logs", "--no-color", "--tail", "200"])
+            .current_dir(directory)
+            .stdin(Stdio::null()),
+        crate::process::SHORT_TIMEOUT,
+        "container logs",
+    )?;
+    if !output.status.success() {
+        return Err(crate::security::environment_error(
+            app,
+            id,
+            String::from_utf8_lossy(&output.stderr).trim(),
+        ));
+    }
+    Ok(crate::security::environment_error(
+        app,
+        id,
+        &format!(
+            "{}{}",
+            String::from_utf8_lossy(&output.stdout),
+            String::from_utf8_lossy(&output.stderr)
+        ),
+    ))
+}
